LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-transition-name-expected.txt

Anchor1
Anchor2

FAIL Transition when the dereferenced anchor name changes assert_equals: expected 0 but got 140
